extract pod_api and the header identification to pay the line ratchet
CI caught this, not me: crates/nucleus-node/src/main.rs has a line
ratchet at 4042 and this branch pushed it to 4081. Two checks went red,
and only one of them was really about lines --

  Check line ceiling            VIOLATION: 4081 > 4042
  Gates must fail on their own  check-line-ratchet.sh still failing
                                after restore

The second is the meta-gate working exactly as intended. It probes each
gate by perturbing its subject and restoring it; a gate that is red before
and after the restore is red on everything, and a gate that always fails
detects nothing. It was reporting that the line ratchet had entered a
permanently-red state, which is a real thing to know and not a duplicate
of the first failure.

Raising the ceiling was not an option -- a ratchet you raise when you hit
it is a comment. The ratchet's own instruction is to extract, and its
target for this file is 2000, so extraction is the sanctioned move.

Two things came out:

* `pod_caller_identity::identify_from_headers` -- the middleware's
  identification block, which belongs beside the derivation and
  verification it calls rather than inline in a request handler.
* `pod_api` -- list_pods / pod_logs / cancel_pod / collect_pod_infos /
  get_pod. The entire read-and-cancel API over NodeState.pods, and the
  exact surface the node-side ownership work extends next, so that change
  now edits a 62-line module instead of growing this file further.

No behaviour changes in either move. 4081 -> 4020, under the ceiling with
22 lines of room.
